spark/host_agent_adk/pyproject.toml
[project]
name = "spark-host-agent"
version = "0.1.0"
description = "SPARK Host Agent - AI-powered digital banking support for transaction discrepancy resolution"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # ADK & A2A Dependencies
    "google-adk>=1.2.1",
    "a2a-sdk>=0.2.5",
    "nest-asyncio>=1.6.0",
    "python-dotenv",
    "click",
    "uvicorn",
    "google-generativeai",
    "httpx",
    
    # Database Dependencies
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    
    # Additional Dependencies
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
]